Refactor: Put shared fixtures into common `conftest.py`
