+++ src/users/tests.py
@@ -13,19 +13,10 @@ faker = Faker()
 User = get_user_model()
 
 
-@pytest.fixture
-def user():
-    return {
-        "email": faker.email(),
-        "password": faker.password(),
-        "display_name": faker.name_nonbinary(),
-        "last_name": faker.last_name(),
-    }
-
-
 @pytest.mark.django_db
 class TestUsersManager:
-    def test_create_user(self, user):
+    def test_create_user(self, signup_user):
+        user = signup_user
         u = User.objects.create_user(email=user["email"], password=user["password"])  # type: ignore
         assert u.email == user["email"]
         assert u.is_active
@@ -44,7 +35,8 @@ class TestUsersManager:
         with pytest.raises(ValueError):
             User.objects.create_user(email="", password="foo")  # type: ignore
 
-    def test_create_superuser(self, user):
+    def test_create_superuser(self, signup_user):
+        user = signup_user
         admin_user = User.objects.create_superuser(
             email=user["email"], password=user["password"]
         )  # type: ignore
@@ -66,13 +58,14 @@ class TestUsersManager:
 
 @pytest.mark.django_db
 class TestUserSignUp:
-    def test_signup_page_has_form(self, client, user):
+    def test_signup_page_has_form(self, client):
         response = cast(HttpResponse, client.get(reverse("users:signup")))
         assert response.status_code == 200
         assert "form" in response.context
         assert isinstance(response.context["form"], CustomUserCreationForm)
 
-    def test_user_can_sign_up(self, client, user):
+    def test_user_can_sign_up(self, client, signup_user):
+        user = signup_user
         response = cast(
             HttpResponse,
             client.post(
@@ -93,11 +86,6 @@ class TestUserSignUp:
         assertContains(response, "Logout")
 
     def test_cannot_signup_duplicate_email(self, client, user):
-        User.objects.create_user(
-            email=user["email"],
-            password=user["password"],
-            display_name=user["display_name"],
-        )  # type: ignore
         response = cast(
             HttpResponse,
             client.post(
@@ -117,13 +105,6 @@ class TestUserSignUp:
 @pytest.mark.django_db
 class TestUserLogin:
     def test_user_can_login(self, client, user):
-        u = User.objects.create_user(
-            email=user["email"],
-            password=user["password"],
-            display_name=user["display_name"],
-            first_name=user["display_name"],
-            last_name=user["last_name"],
-        )  # type: ignore
         response = cast(
             HttpResponse,
             client.post(
@@ -136,7 +117,7 @@ class TestUserLogin:
             ),
         )
         assert response.status_code == 200
-        assert response.context["user"].email == u.email
+        assert response.context["user"].email == user["email"]
         last_url, _ = response.redirect_chain[-1]
         assert last_url == reverse("dashboard:index")
 
@@ -146,8 +127,8 @@ class TestUserLogin:
             client.post(
                 reverse("users:login"),
                 data={
-                    "email": user["email"],
-                    "password": user["password"],
+                    "email": faker.email(),
+                    "password": faker.password(),
                 },
             ),
         )
